requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-groq
chromadb
pypdf
python-dotenv
sentence-transformers
streamlit
watchdog
